Adds the Object IDs type to the exporter.
