+++ lib/bulkrax/override_assistive_methods.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+require_relative './export_assistive_methods'
+require_relative './ingest_assistive_methods'
+
+module OverrideAssistiveMethods
+  include ExportAssistiveMethods
+  include IngestAssistiveMethods
+end
